update reverse mode functionality
